Fixed Database, Fixed LogIn and BugFixes
Moved to PostgreSQL (on https://pgserver.mau.se) from the previous group's Azure.
Created the ability to check if tables  (user, plant) exist, otherwise create them. This was based on the user and plant classes and repositories.
Fixed the reference to the StyleSheet.css file in the various Java files.
Fixed various small bugs.
The app now is able to connect to my postgresql database, to both myhappyplant and species database names and the login functionality works, sending the user to the main page after a successful login.
Added PasswordsAndKeys.java into .gitignore file. THIS IS A SENSITIVE FILE!